requirements.txt
google-genai
moviepy
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pillow
